+++ cfn-resources/network-container/cmd/resource/resource.go
@@ -15,277 +15,35 @@
 package resource
 
 import (
-	"context"
-	"fmt"
-	"log"
-	"net/http"
-
 	"github.com/aws-cloudformation/cloudformation-cli-go-plugin/cfn/handler"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/cloudformation"
-	"github.com/mongodb/mongodbatlas-cloudformation-resources/profile"
-	"github.com/mongodb/mongodbatlas-cloudformation-resources/util"
-	"github.com/mongodb/mongodbatlas-cloudformation-resources/util/constants"
-	"github.com/mongodb/mongodbatlas-cloudformation-resources/util/logger"
-	progressevents "github.com/mongodb/mongodbatlas-cloudformation-resources/util/progressevent"
 	"github.com/mongodb/mongodbatlas-cloudformation-resources/util/validator"
-	"go.mongodb.org/atlas/mongodbatlas"
 )
-
-var CreateRequiredFields = []string{constants.ProjectID, constants.RegionName, constants.AtlasCIDRBlock}
-var ReadRequiredFields = []string{constants.ProjectID, constants.ID}
-var UpdateRequiredFields = []string{constants.ProjectID, constants.ID}
-var DeleteRequiredFields = []string{constants.ProjectID, constants.ID}
-var ListRequiredFields = []string{constants.ProjectID}
-
-// function to validate inputs to all actions
-func validateModel(fields []string, model *Model) *handler.ProgressEvent {
-	return validator.ValidateModel(fields, model)
-}
 
 // Create handles the Create event from the Cloudformation service.
 func Create(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
-	if errEvent := validateModel(CreateRequiredFields, currentModel); errEvent != nil {
-		return *errEvent, nil
-	}
-
-	// Create atlas client
-	if currentModel.Profile == nil || *currentModel.Profile == "" {
-		currentModel.Profile = aws.String(profile.DefaultProfile)
-	}
-	client, peErr := util.NewMongoDBClient(req, currentModel.Profile)
-	if peErr != nil {
-		return *peErr, nil
-	}
-
-	projectID := currentModel.ProjectId
-
-	containerRequest := &mongodbatlas.Container{}
-
-	if projectID == nil || *projectID == "" {
-		return handler.ProgressEvent{}, fmt.Errorf("error creating network container: `%s` must be set", constants.ProjectID)
-	}
-
-	regionName := currentModel.RegionName
-	if regionName == nil || *regionName == "" {
-		return handler.ProgressEvent{}, fmt.Errorf("`error creating network container: `%s` must be set", constants.RegionName)
-	}
-	containerRequest.RegionName = *regionName
-	containerRequest.ProviderName = constants.AWS
-	CIDR := currentModel.AtlasCidrBlock
-	if CIDR == nil || *CIDR == "" {
-		return handler.ProgressEvent{}, fmt.Errorf("error creating network container: `%s` must be set", constants.AtlasCIDRBlock)
-	}
-	containerRequest.AtlasCIDRBlock = *CIDR
-	containerResponse, res, err := client.Containers.Create(context.Background(), *projectID, containerRequest)
-	if err != nil {
-		if res.StatusCode == http.StatusConflict {
-			_, _ = logger.Debugf("Container already exists for this group. Try return existing container. err: %v", err)
-			containers, _, err := client.Containers.ListAll(context.Background(), *projectID, nil)
-			if err != nil {
-				_, _ = logger.Debugf("Error Containers.ListAll err:%v", err)
-				return handler.ProgressEvent{}, fmt.Errorf("error Containers.ListAll err:%v", err)
-			}
-
-			for i := range containers {
-				if containers[i].RegionName == *regionName {
-					currentModel.Id = &containers[i].ID
-					return handler.ProgressEvent{
-						OperationStatus: handler.Success,
-						Message:         "Create complete",
-						ResourceModel:   currentModel,
-					}, nil
-				}
-			}
-
-			return handler.ProgressEvent{}, fmt.Errorf("error creating network container: %s", err)
-		}
-
-		return handler.ProgressEvent{}, fmt.Errorf("error creating network container: %s", err)
-	}
-
-	currentModel.Id = &containerResponse.ID
-
-	_, _ = logger.Debugf("Create about to return this --->> currentModel:%+v", currentModel)
-	return handler.ProgressEvent{
-		OperationStatus: handler.Success,
-		Message:         "Create complete",
-		ResourceModel:   currentModel,
-	}, nil
+	return CreateOperation(req, prevModel, currentModel)
 }
 
 // Read handles the Read event from the Cloudformation service.
 func Read(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
-	if errEvent := validateModel(ReadRequiredFields, currentModel); errEvent != nil {
-		return *errEvent, nil
-	}
-
-	// Create atlas client
-	if currentModel.Profile == nil || *currentModel.Profile == "" {
-		currentModel.Profile = aws.String(profile.DefaultProfile)
-	}
-	client, peErr := util.NewMongoDBClient(req, currentModel.Profile)
-	if peErr != nil {
-		return *peErr, nil
-	}
-
-	projectID := *currentModel.ProjectId
-	containerID := *currentModel.Id
-
-	containerResponse, response, err := client.Containers.Get(context.Background(), projectID, containerID)
-
-	if err != nil {
-		return progressevents.GetFailedEventByResponse(fmt.Sprintf("Error getting resource : %s", err.Error()),
-			response.Response), nil
-	}
-
-	currentModel.RegionName = &containerResponse.RegionName
-	currentModel.Provisioned = containerResponse.Provisioned
-	currentModel.VpcId = &containerResponse.VPCID
-	currentModel.AtlasCidrBlock = &containerResponse.AtlasCIDRBlock
-
-	return handler.ProgressEvent{
-		OperationStatus: handler.Success,
-		Message:         "Read Complete",
-		ResourceModel:   currentModel,
-	}, nil
+	return ReadOperation(req, prevModel, currentModel)
 }
 
 // Update handles the Update event from the Cloudformation service.
 func Update(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
-	if errEvent := validateModel(UpdateRequiredFields, currentModel); errEvent != nil {
-		return *errEvent, nil
-	}
-
-	// Create atlas client
-	if currentModel.Profile == nil || *currentModel.Profile == "" {
-		currentModel.Profile = aws.String(profile.DefaultProfile)
-	}
-	client, peErr := util.NewMongoDBClient(req, currentModel.Profile)
-	if peErr != nil {
-		return *peErr, nil
-	}
-
-	projectID := *currentModel.ProjectId
-	containerID := *currentModel.Id
-	containerRequest := &mongodbatlas.Container{}
-
-	CIDR := currentModel.AtlasCidrBlock
-	if CIDR != nil {
-		containerRequest.AtlasCIDRBlock = *CIDR
-	}
-	containerRequest.ProviderName = constants.AWS
-	containerRequest.RegionName = *currentModel.RegionName
-	containerResponse, resp, err := client.Containers.Update(context.Background(), projectID, containerID, containerRequest)
-	if err != nil {
-		return progressevents.GetFailedEventByResponse(fmt.Sprintf("Error getting resource : %s", err.Error()),
-			resp.Response), nil
-	}
-
-	currentModel.Id = &containerResponse.ID
-	_, _ = logger.Debugf("Create network container - Id: %v", currentModel.Id)
-
-	return handler.ProgressEvent{
-		OperationStatus: handler.Success,
-		Message:         "Update Complete",
-		ResourceModel:   currentModel,
-	}, nil
+	return UpdateOperation(req, prevModel, currentModel)
 }
 
 // Delete handles the Delete event from the Cloudformation service.
 func Delete(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
-	_, _ = logger.Debugf("Delete currentModel:%+v", currentModel)
-
-	if errEvent := validateModel(DeleteRequiredFields, currentModel); errEvent != nil {
-		return *errEvent, nil
-	}
-
-	// Create atlas client
-	if currentModel.Profile == nil || *currentModel.Profile == "" {
-		currentModel.Profile = aws.String(profile.DefaultProfile)
-	}
-	client, peErr := util.NewMongoDBClient(req, currentModel.Profile)
-	if peErr != nil {
-		return *peErr, nil
-	}
-
-	_, _ = logger.Debugf("Delete currentModel:%+v", currentModel)
-	projectID := *currentModel.ProjectId
-	containerID := *currentModel.Id
-
-	if response, err := client.Containers.Delete(context.Background(), projectID, containerID); err != nil {
-		if response.StatusCode == 409 {
-			return handler.ProgressEvent{
-				OperationStatus:  handler.Failed,
-				Message:          fmt.Sprintf("Please, make sure to delete the network peering and the atlas cluster before deleting the container: %s", err.Error()),
-				HandlerErrorCode: cloudformation.HandlerErrorCodeResourceConflict,
-			}, nil
-		}
-
-		return progressevents.GetFailedEventByResponse(fmt.Sprintf("Error getting resource: %s", err.Error()),
-			response.Response), nil
-	}
-
-	return handler.ProgressEvent{
-		OperationStatus: handler.Success,
-		Message:         "Delete Complete",
-	}, nil
+	return DeleteOperation(req, prevModel, currentModel)
 }
 
 // List handles the List event from the Cloudformation service.
 func List(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
-	_, _ = logger.Debugf("List currentModel:%+v", currentModel)
-	log.SetFlags(log.LstdFlags | log.Lshortfile)
-
-	_, _ = logger.Debugf("List currentModel:%+v", currentModel)
-
-	if errEvent := validateModel(ListRequiredFields, currentModel); errEvent != nil {
-		return *errEvent, nil
-	}
-
-	// Create atlas client
-	if currentModel.Profile == nil || *currentModel.Profile == "" {
-		currentModel.Profile = aws.String(profile.DefaultProfile)
-	}
-	client, peErr := util.NewMongoDBClient(req, currentModel.Profile)
-	if peErr != nil {
-		return *peErr, nil
-	}
-
-	projectID := *currentModel.ProjectId
-	containerRequest := &mongodbatlas.ContainersListOptions{
-		ProviderName: constants.AWS,
-		ListOptions:  mongodbatlas.ListOptions{},
-	}
-	_, _ = logger.Debugf("List projectId:%v, containerRequest:%v", projectID, containerRequest)
-	containerResponse, _, err := client.Containers.List(context.TODO(), projectID, containerRequest)
-	if err != nil {
-		_, _ = logger.Warnf("Error %v", err)
-		return handler.ProgressEvent{}, err
-	}
-
-	_, _ = logger.Debugf("containerResponse:%v", containerResponse)
-
-	mm := make([]interface{}, 0)
-	for i := range containerResponse {
-		var m Model
-		m.completeByConnection(containerResponse[i])
-		m.ProjectId = currentModel.ProjectId
-		m.Profile = currentModel.Profile
-		mm = append(mm, m)
-	}
-
-	return handler.ProgressEvent{
-		OperationStatus: handler.Success,
-		Message:         "List Complete",
-		ResourceModels:  mm,
-	}, nil
+	return ListOperation(req, prevModel, currentModel)
 }
 
-func (m *Model) completeByConnection(c mongodbatlas.Container) {
-	m.RegionName = &c.RegionName
-	m.Provisioned = c.Provisioned
-	m.Id = &c.ID
-	m.VpcId = &c.VPCID
-	m.AtlasCidrBlock = &c.AtlasCIDRBlock
+func validateModel(fields []string, model *Model) *handler.ProgressEvent {
+	return validator.ValidateModel(fields, model)
 }
