+++ cfn-resources/network-container/cmd/resource/update.go
@@ -1,0 +1,71 @@
+// Copyright 2023 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package resource
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/aws-cloudformation/cloudformation-cli-go-plugin/cfn/handler"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/mongodb/mongodbatlas-cloudformation-resources/profile"
+	"github.com/mongodb/mongodbatlas-cloudformation-resources/util"
+	"github.com/mongodb/mongodbatlas-cloudformation-resources/util/constants"
+	"github.com/mongodb/mongodbatlas-cloudformation-resources/util/logger"
+	progressevents "github.com/mongodb/mongodbatlas-cloudformation-resources/util/progressevent"
+	"go.mongodb.org/atlas/mongodbatlas"
+)
+
+var UpdateRequiredFields = []string{constants.ProjectID, constants.ID}
+
+func UpdateOperation(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
+	if errEvent := validateModel(UpdateRequiredFields, currentModel); errEvent != nil {
+		return *errEvent, nil
+	}
+
+	// Create atlas client
+	if currentModel.Profile == nil || *currentModel.Profile == "" {
+		currentModel.Profile = aws.String(profile.DefaultProfile)
+	}
+	client, peErr := util.NewMongoDBClient(req, currentModel.Profile)
+	if peErr != nil {
+		return *peErr, nil
+	}
+
+	projectID := *currentModel.ProjectId
+	containerID := *currentModel.Id
+	containerRequest := &mongodbatlas.Container{}
+
+	CIDR := currentModel.AtlasCidrBlock
+	if CIDR != nil {
+		containerRequest.AtlasCIDRBlock = *CIDR
+	}
+	containerRequest.ProviderName = constants.AWS
+	containerRequest.RegionName = *currentModel.RegionName
+	containerResponse, resp, err := client.Containers.Update(context.Background(), projectID, containerID, containerRequest)
+	if err != nil {
+		return progressevents.GetFailedEventByResponse(fmt.Sprintf("Error getting resource : %s", err.Error()),
+			resp.Response), nil
+	}
+
+	currentModel.Id = &containerResponse.ID
+	_, _ = logger.Debugf("Create network container - Id: %v", currentModel.Id)
+
+	return handler.ProgressEvent{
+		OperationStatus: handler.Success,
+		Message:         "Update Complete",
+		ResourceModel:   currentModel,
+	}, nil
+}
